+++ components/form-builder/panel/index.ts
@@ -1,1 +1,9 @@
 export { ElementPanel } from "./ElementPanel";
+export { PanelActions } from "./PanelActions";
+export { PanelActionsLocked } from "./PanelActionsLocked";
+export { QuestionInput } from "./QuestionInput";
+export { ConfirmationDescription } from "./ConfirmationDescription";
+export { Modal, ModalButton } from "./Modal";
+export { ModalForm } from "./ModalForm";
+export { PrivacyDescription } from "./PrivacyDescription";
+export { SelectedElement, getSelectedOption } from "./SelectedElement";
