+++ components/form-builder/shared/index.ts
@@ -1,0 +1,6 @@
+export { Button } from "./Button";
+export { ConfirmFormDeleteDialog } from "./ConfirmFormDeleteDialog";
+export { Dialog } from "./Dialog";
+export { Input } from "./Input";
+export { TextArea } from "./TextArea";
+export { Radio, Checkbox } from "./MultipleChoice";
